add events of different sizes
